Moved BibTeX processing from OrcidWork into Work, to avoid processing when doing visibility filtering (and hence bogging down the scheduler).
